+++ source/module_esolver/esolver_ks_lcao_elec.cpp
@@ -8,8 +8,8 @@
 #include "module_cell/module_neighbor/sltk_atom_arrange.h"
 #include "module_cell/module_neighbor/sltk_grid_driver.h"
 #include "module_io/berryphase.h"
-#include "module_io/istate_charge.h"
-#include "module_io/istate_envelope.h"
+#include "module_io/get_pchg.h"
+#include "module_io/get_wf.h"
 #include "module_io/to_wannier90_lcao.h"
 #include "module_io/to_wannier90_lcao_in_pw.h"
 #include "module_io/write_HS_R.h"
