+++ source/module_io/CMakeLists.txt
@@ -50,8 +50,8 @@ if(ENABLE_LCAO)
       write_orb_info.cpp
       write_proj_band_lcao.cpp
       nscf_fermi_surf.cpp
-      istate_charge.cpp
-      istate_envelope.cpp
+      get_pchg.cpp
+      get_wf.cpp
       read_wfc_nao.cpp
       read_wfc_lcao.cpp
       write_wfc_nao.cpp
